config: wipe installation settings when the directory is changed
